Added devices classes to various sensors.
Various sensors were missing device classes.
It would seem this is a requirement for built in unit conversion in Home Assistant.  By adding speed and distance classes, Home Assistant is able to provide conversion to miles and mph without the need to set up template sensors to do this for us.